1cnn/nodejs-backend/requirements.txt
tensorflow>=2.10.0
flask>=2.0.0
flask-cors>=3.0.10
Pillow>=9.0.0
numpy>=1.20.0 